requirements.txt
Flask==2.0.1
pdf2image==1.15.1
Pillow==9.3
requests==2.26.0
gunicorn==19.9.0
PyPDF2==1.26.0
PyMuPDF==1.18.14